Introduce a metabox for capturing event cost data
This is most likely woefully inadequate. Price often depends on
the customer (student, faculty/staff, child, general public), so
we'll need to determine if we can reliably account for every case
width static inputs - or on the flip side, provide repeatable
inputs with which the user can define customer type and price.